Conflicts:
	shotgun-client/android/.gradle/2.14.1/taskArtifacts/cache.properties.lock
	shotgun-client/android/.gradle/2.14.1/taskArtifacts/fileHashes.bin
	shotgun-client/android/.gradle/2.14.1/taskArtifacts/fileSnapshots.bin
	shotgun-client/android/.gradle/2.14.1/taskArtifacts/fileSnapshotsToTreeSnapshotsIndex.bin
	shotgun-client/android/.gradle/2.14.1/taskArtifacts/taskArtifacts.bin
	shotgun-client/app/common/components/PagingListView.js
	shotgun-client/app/customer/product/ProductCategoryList.js
	shotgun-client/app/customer/registration/CustomerDetails.js
	shotgun-client/package.json
	viewserver-shotgun/src/main/resources/data/product.csv

+++ viewserver-shotgun/src/main/java/com/shotgun/viewserver/setup/datasource/ProductDataSource.java
@@ -23,8 +23,8 @@ ProductDataSource {
                         new Column("name", "name", ColumnType.String),
                         new Column("description", "description", ColumnType.String),
                         new Column("categoryId", "categoryId", ColumnType.String),
-                        new Column("price", "price", ColumnType.Double),
-                        new Column("rating", "rating", ColumnType.Double)
+                        new Column("price", "price", ColumnType.Int),
+                        new Column("rating", "rating", ColumnType.Int)
                 ))
                 .withKeyColumns("productId");
